requirements.txt
# requirements.txt
google-generativeai
python-dotenv~=1.1.0
requests~=2.32.3
beautifulsoup4~=4.13.3
protobuf~=5.29.4